move normal into smudging
